refactor(8.10): drop orphan helpers and guard keycloak release info
