+++ charts/camunda-platform-8.10/README.md
@@ -721,7 +721,7 @@ Please see the corresponding [release guide](../../docs/release.md) to find out 
 | `webModeler.restapi.image.digest`                                      | can be used to set image digest (overrides tag if set, e.g. "sha256:abcd...")                                                                                                                        | `""`                          |
 | `webModeler.restapi.sidecars`                                          | can be used to attach extra containers to the restapi deployment                                                                                                                                     | `[]`                          |
 | `webModeler.restapi.initContainers`                                    | can be used to set up extra init containers for the application Pod                                                                                                                                  | `[]`                          |
-| `webModeler.restapi.externalDatabase`                                  | can be used to configure a connection to an external database; will only be applied if the postgresql dependency chart is disabled (with `postgresql.enabled=false`)                                 |                               |
+| `webModeler.restapi.externalDatabase`                                  | can be used to configure a connection to an external database for Web Modeler                                                                                                                        |                               |
 | `webModeler.restapi.externalDatabase.url`                              | defines the JDBC url of the database instance.                                                                                                                                                       | `""`                          |
 | `webModeler.restapi.externalDatabase.host`                             | Database host (used when url is not provided)                                                                                                                                                        | `""`                          |
 | `webModeler.restapi.externalDatabase.port`                             | Database port number (used when url is not provided)                                                                                                                                                 | `5432`                        |
