--- charts/camunda-platform-8.10/templates/z/1/2/3/4/5/6/7/8/z_compatibility_helpers.tpl
@@ -1,34 +1,0 @@
-{{/*
-NOTE: We need to load this file first thing before all other resources to support backward compatibility.
-
-      Helm prioritizes files that are deeply nested in subdirectories when it is determining the render order.
-      see the sort function in Helm:
-      https://github.com/helm/helm/blob/d58d7b376265338e059ff11c71267b5a6cf504c3/pkg/engine/engine.go#L347-L356
-
-      Because of this sort order, and that we have nested subcharts such that
-      one of the rendered templates is:
-      charts/keycloak/charts/postgresql/charts/common/templates/validations/_validations.tpl,
-      we need this z_compatibility_helpers.tpl to be nested in at least 8 folders.
-
-      In addition to the subdirectory ordering, Helm also orders the templates
-      alphabetically descending within the same folder level, which is why it
-      is named with a "z_" inside the zeebe directory. so Helm will process
-      this file first, and all migration steps will be applied to all templates
-      in the chart implicitly:
-      https://github.com/helm/helm/blob/d58d7b376265338e059ff11c71267b5a6cf504c3/pkg/engine/engine.go#L362-L369
-*/}}
-
-
-{{/*
-********************************************************************************
-* Camunda 8.5 backward compatibility.
-
-Overview:
-    Backward compatibility with values syntax before Camunda Helm chart v10.0.0 (Camunda 8.5 cycle).
-
-Approach:
-    Using deep copy and deep merge functions to override new keys using the old key.
-    https://helm.sh/docs/chart_template_guide/function_list/#mergeoverwrite-mustmergeoverwrite
-********************************************************************************
-*/}}
-
